修改代码

+++ src/main/java/org/athena/utils/Constant.java
@@ -7,7 +7,7 @@ public final class Constant {
 
     public static final String AUTHORIZATION_HEADER = "Authorization";
 
-    public static final Long AUTHORIZATION_DURATION = 120L;
+    public static final Integer AUTHORIZATION_DURATION = 120;
 
     public static final String TIME_ZONE = "UTC";
